Added VersionController for git version functions
